test(lyrics): move new lyrics tests into master's test subdirs
Master reorganized tests into per-area subdirs (test/db, test/scanner,
test/unit, test/integration, test/subsonic) with glob-based npm scripts.
The lyrics PR's new test files were still at the top level, so `npm test`
(recursive glob) ran them but the categorized scripts (test:db etc.) did
not. Move them to match, fixing the relative import/path depth:

  admin-lyrics-settings        -> test/integration/
  lyrics-backfill              -> test/integration/  (mirrors album-art-downloader)
  db-fts5-lyrics               -> test/db/
  lyrics-lookup-lib            -> test/unit/
  scanner-lyrics-clobber-guard -> test/scanner/
